+++ src/components/manage/BlockChooser/BlockChooser.jsx
@@ -29,10 +29,23 @@ const BlockChooser = ({
   blocksConfig = config.blocks.blocksConfig,
   intl,
 }) => {
-  const filteredBlocksConfig = filter(
-    blocksConfig,
-    (item) => isEmpty(allowedBlocks) || allowedBlocks.includes(item.id),
-  );
+  const useAllowedBlocks = !isEmpty(allowedBlocks);
+
+  const filteredBlocksConfig = filter(blocksConfig, (item) => {
+    if (showRestricted) {
+      if (useAllowedBlocks) {
+        return allowedBlocks.includes(item.id);
+      } else {
+        return true;
+      }
+    } else {
+      if (useAllowedBlocks) {
+        return allowedBlocks.includes(item.id);
+      } else {
+        return !item.restricted;
+      }
+    }
+  });
 
   let blocksAvailable = {};
   const mostUsedBlocks = filter(filteredBlocksConfig, (item) => item.mostUsed);
@@ -97,30 +110,24 @@ const BlockChooser = ({
                 duration={500}
                 height={activeIndex === index ? 'auto' : 0}
               >
-                {map(
-                  filter(
-                    blocksAvailable[groupName.id],
-                    (block) => showRestricted || !block.restricted,
-                  ),
-                  (block) => (
-                    <Button.Group key={block.id}>
-                      <Button
-                        icon
-                        basic
-                        className={block.id}
-                        onClick={() =>
-                          onMutateBlock(currentBlock, { '@type': block.id })
-                        }
-                      >
-                        <Icon name={block.icon} size="36px" />
-                        {intl.formatMessage({
-                          id: block.id,
-                          defaultMessage: block.title,
-                        })}
-                      </Button>
-                    </Button.Group>
-                  ),
-                )}
+                {map(blocksAvailable[groupName.id], (block) => (
+                  <Button.Group key={block.id}>
+                    <Button
+                      icon
+                      basic
+                      className={block.id}
+                      onClick={() =>
+                        onMutateBlock(currentBlock, { '@type': block.id })
+                      }
+                    >
+                      <Icon name={block.icon} size="36px" />
+                      {intl.formatMessage({
+                        id: block.id,
+                        defaultMessage: block.title,
+                      })}
+                    </Button>
+                  </Button.Group>
+                ))}
               </AnimateHeight>
             </Accordion.Content>
           </React.Fragment>