+++ src/components/manage/BlockChooser/BlockChooser.test.jsx
@@ -140,6 +140,19 @@ describe('BlocksChooser', () => {
     // There are 2 because the others are aria-hidden="true"
     expect(screen.getAllByRole('button')).toHaveLength(2);
   });
+  it('allowedBlocks bypasses showRestricted', () => {
+    config.blocks.blocksConfig.listing.restricted = true;
+    const { container } = render(
+      <Provider store={store}>
+        <BlockChooser
+          onMutateBlock={() => {}}
+          currentBlock="theblockid"
+          allowedBlocks={['image', 'listing']}
+        />
+      </Provider>,
+    );
+    expect(container).toHaveTextContent('Listing');
+  });
   it('showRestricted test', () => {
     const { container } = render(
       <Provider store={store}>
@@ -181,6 +194,40 @@ describe('BlocksChooser', () => {
         />
       </Provider>,
     );
-    expect(container.firstChild).toHaveTextContent('Custom block');
+    expect(container).toHaveTextContent('Custom block');
+  });
+  it("doesn't show empty groups", () => {
+    const old = [...config.blocks.groupBlocksOrder];
+    config.blocks.groupBlocksOrder.push({
+      id: 'customGroup',
+      title: 'Custom group',
+    });
+    const blocksConfig = {
+      ...config.blocks.blocksConfig,
+      custom: {
+        id: 'custom',
+        title: 'Custom block',
+        icon: blockSVG,
+        mostUsed: true,
+        group: 'customGroup',
+        restricted: true,
+        edit: ({ id, data }) => (
+          <div>
+            {id} - {data.text}
+          </div>
+        ),
+      },
+    };
+    const { container } = render(
+      <Provider store={store}>
+        <BlockChooser
+          onMutateBlock={() => {}}
+          currentBlock="theblockid"
+          blocksConfig={blocksConfig}
+        />
+      </Provider>,
+    );
+    expect(container).not.toHaveTextContent('Custom group');
+    config.blocks.groupBlocksOrder = old;
   });
 });